+++ packages/expo-sqlite/android/src/main/cpp/NativeStatementBinding.cpp
@@ -32,6 +32,7 @@ void NativeStatementBinding::registerNatives() {
                        NativeStatementBinding::sqlite3_finalize),
       makeNativeMethod("sqlite3_reset", NativeStatementBinding::sqlite3_reset),
       makeNativeMethod("sqlite3_step", NativeStatementBinding::sqlite3_step),
+      makeNativeMethod("isNoop", NativeStatementBinding::isNoop),
       makeNativeMethod("bindStatementParam",
                        NativeStatementBinding::bindStatementParam),
       makeNativeMethod("getColumnNames",
@@ -65,6 +66,8 @@ int NativeStatementBinding::sqlite3_finalize() {
 int NativeStatementBinding::sqlite3_reset() { return ::exsqlite3_reset(stmt); }
 
 int NativeStatementBinding::sqlite3_step() { return ::exsqlite3_step(stmt); }
+
+bool NativeStatementBinding::isNoop() { return stmt == nullptr; }
 
 int NativeStatementBinding::bindStatementParam(
     int index, jni::alias_ref<jni::JObject> param) {